+++ R/sys.R
@@ -1,4 +1,4 @@
 .tmax <- 100L
 .Lmax <- 1e+04
 .parameters <- c("tmax", "k", "Linf", "t0", "b", "Lm", "fb", "Rt",
-                "nu", "Lv", "Llo", "Lup", "mu", "rho", "eta", "Rk")
+                "nu", "Lv", "Vp", "Llo", "Lup", "mu", "rho", "eta", "Rk")