+++ NEWS.md
@@ -1,5 +1,6 @@
 - Added `ypr_yield_uncertainty()`, `ypr_yields_uncertainty()` and 
 `ypr_optimize_uncertainty()` to account for uncertainty in parameters.
+- Added vulnerability power `Vp`.
 - Increased upper limit for `Rk` from 10 to 100.
 - Removed `a` and `fa` as just scaling terms.